feat : #FINAL-21 객실 서비스 수정
- feign response 수정
- feign 패키지 경로 수정
- room service와 accommodation service 통신 수정
