+++ reservation-service/src/main/java/com/fp/reservationservice/global/feign/dto/response/room/RoomImageResponse.java
@@ -3,9 +3,11 @@ package com.fp.reservationservice.global.feign.dto.response.room;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @AllArgsConstructor
+@NoArgsConstructor
 public class RoomImageResponse {
 
     private List<String> imageUrlList;
