+++ room-service/src/main/java/com/fp/roomservice/domain/dto/response/RoomSimpleResponse.java
@@ -1,0 +1,39 @@
+package com.fp.roomservice.domain.dto.response;
+
+
+import com.fp.roomservice.domain.entity.Room;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+@AllArgsConstructor
+public class RoomSimpleResponse {
+
+    private Long id;
+
+    private String name;
+
+    private String checkInTime;
+
+    private String checkOutTime;
+
+    private int standardNumber;
+
+    private int maximumNumber;
+
+    private RoomImageResponse imageResponse;
+
+    public static RoomSimpleResponse from(Room room, RoomImageResponse imageResponse) {
+        return RoomSimpleResponse.builder()
+            .id(room.getId())
+            .name(room.getName())
+            .checkInTime(room.getCheckInTime())
+            .checkOutTime(room.getCheckOutTime())
+            .standardNumber(room.getStandardNumber())
+            .maximumNumber(room.getMaximumNumber())
+            .imageResponse(imageResponse)
+            .build();
+    }
+}
