--- room-service/src/main/java/com/fp/roomservice/entity/RoomInfo.java
@@ -1,37 +1,0 @@
-package com.fp.roomservice.entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import java.time.LocalDate;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RoomInfo {
-
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Long roomId;
-
-    private LocalDate date;
-
-    private int price;
-
-    private int count;
-
-    public void decreaseCountByOne() {
-        this.count--;
-    }
-}
