--- room-service/src/main/java/com/fp/roomservice/exception/RoomException.java
@@ -1,15 +1,0 @@
-package com.fp.roomservice.exception;
-
-import org.springframework.web.client.HttpStatusCodeException;
-
-public class RoomException extends HttpStatusCodeException {
-
-    public RoomException(ErrorType errorType) {
-        super(errorType.getStatusCode(), errorType.getMessage());
-    }
-
-    @Override
-    public String getMessage() {
-        return getStatusText();
-    }
-}
