--- room-service/src/main/java/com/fp/roomservice/repository/RoomImageRepository.java
@@ -1,10 +1,0 @@
-package com.fp.roomservice.repository;
-
-import com.fp.roomservice.entity.RoomImage;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RoomImageRepository extends JpaRepository<RoomImage, Long> {
-
-    List<RoomImage> findAllByRoomId(Long roomId);
-}
